Port outfit-derived state arms and mission payload context
Mission scripts:
- Mission_RunMisnScriptPayload now sets/clears the mission context slot
  around the engine so the Q opcode expands mission text tags, recomputes
  outfit-derived state, and matches the original's lack of slot validation
  (nested S-payloads clobber the context, as the original global does).
- Mission_ExecuteReactionScript gains the empty early-return and recompute.

Outfit-derived state (0x0046d4b0):
- Rename OutfitMarkStatsDirty -> NovaOutfit_RecomputeOutfitDerivedState,
  internal helper -> InvalidatePlayerStatCache.
- Port negative cargo/junk clamps, player cloak-latch reset, recently-hit
  timer reset, and the owned ModType 0x2c/0x30 government latches
  (GovtDef +0x82 reinforcement_inhibited / +0x83 iff_scrambler_active,
  g_player_reinforcement_inhibit_all).
- Port the policy_flags clear/rebuild from active ranks; this is the
  writer Government_GetPolicyFlag consumers were missing.

Naming:
- Ghidra: g_script_mission_context_slot, g_pending_overlay_message,
  g_pending_transient_sound_id, Resource_LoadRandomStringEntry.
- GameState: script_mission_context_slot, pending_transient_sound_id,
  pending_script_message_string_id.

Docs/trackers:
- New docs/outfit_derived_state.md (two-mechanism model, caller map,
  remaining arms); scenario_data_loading cross-reference.
- 0x00448020/0x00448050 -> 100%, 0x00449370 -> 80%, 0x0046D4B0 -> 50%.

+++ src/game/mission_script.hpp
@@ -37,7 +37,7 @@ MissionScriptResult Mission_ExecuteReactionScript(GameState &state,
                                                   std::string_view script);
 MissionScriptResult Mission_RunMisnScriptPayload(GameState &state,
                                                  std::string_view script,
-                                                 std::size_t mission_slot);
+                                                 std::int16_t mission_slot);
 MissionScriptResult Mission_ExecuteMisnScriptEngine(GameState &state,
                                                     std::string_view script);
